moved package update thing to a separate view
